% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_featureME.R
\name{get_featureME}
\alias{get_featureME}
\title{Conducts multiple linear regression using item factors and mode effect per item}
\usage{
get_featureME(filename = "")
}
\arguments{
\item{filename}{A file path to a binary csv file that stores features of items as columns. The first column stores item ID as a string.}
}
\value{
Returns the results of the multiple linear regression as a dataframe
}
\description{
Conducts multiple linear regression using item factors and mode effect per item
}
\examples{
get_featureME("featurefile.csv")
}
