% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusters.R
\name{getClusters}
\alias{getClusters}
\title{Returns clustered-items in a dataframe of crossed itemIDs where paired-items with the same label do not show different mode effect}
\usage{
getClusters()
}
\value{
Returns a dataframe that indicates membership to clusters. A negative label indicates that the cluster does not form a complete graph, positive indicates a complete graph, and 0 indicates non-membership to a cluster of more than two items
}
\description{
Returns clustered-items in a dataframe of crossed itemIDs where paired-items with the same label do not show different mode effect
}
\examples{
getClusters()
}
