% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClustItems.R
\name{getClustItems}
\alias{getClustItems}
\title{Gets the ID of the items that have the cluster number that is passed as an argument}
\usage{
getClustItems(groupN = 0)
}
\arguments{
\item{groupN}{One of the cluster numbers. See getClusters()}
}
\value{
Returns a vector of itemIDs as characters
}
\description{
Gets the ID of the items that have the cluster number that is passed as an argument
}
\examples{
getClustItems(2)
}
