% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInvGroup.R
\name{getInvGroup}
\alias{getInvGroup}
\title{Finds the numeric label of the cluster of items that has range of mode effect inclusive of zero}
\usage{
getInvGroup()
}
\value{
The numeric label of the cluster of items that has range of mode effect inclusive of zero
}
\description{
Finds the numeric label of the cluster of items that has range of mode effect inclusive of zero
}
\examples{
getInvGroup()
}
