% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRange.R
\name{getRange}
\alias{getRange}
\title{getRange}
\usage{
getRange(clusterN = -1e+05)
}
\arguments{
\item{clusterN}{The number of the cluster from getClusters()}
}
\value{
Range of mode effect of the given cluster
}
\description{
getRange
}
\examples{
getRange(1)
}
