% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp1.R
\name{hyp1}
\alias{hyp1}
\title{A wrapper function for lfe::waldtest to test the null hypothesis
that the overall mode effect equals 0}
\usage{
hyp1()
}
\value{
1 if overall mode effect is statistically different from zero
}
\description{
A wrapper function for lfe::waldtest to test the null hypothesis
that the overall mode effect equals 0
}
\examples{
hyp1()
}
