% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp2.R
\name{hyp2}
\alias{hyp2}
\title{A wrapper function for one-way ANOVA on mode effect (aov) with item ID as the predictor}
\usage{
hyp2()
}
\value{
Return value of 1 indicates items contribute differently to mode effect (p <= 0.10)
}
\description{
A wrapper function for one-way ANOVA on mode effect (aov) with item ID as the predictor
}
\examples{
hyp2()
}
