% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp3.R
\name{hyp3}
\alias{hyp3}
\title{A wrapper function for Tukey's test (TukeyHSD) for post-hoc, multiple comparisons of item-level mode effect differences}
\usage{
hyp3()
}
\value{
A dataframe showing item-by-item comparisons of mode effect difference
}
\description{
A wrapper function for Tukey's test (TukeyHSD) for post-hoc, multiple comparisons of item-level mode effect differences
}
\examples{
hyp3()
}
