% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemmodef.R
\name{itemmodef}
\alias{itemmodef}
\title{Initializing function that accepts a dataframe with paper and computer based measures for either item-difficulty,
-discrimination, or -guessing.}
\usage{
itemmodef(pbcbDF = data.frame())
}
\arguments{
\item{pbcbDF}{A dataframe with three columns. The first column is item ID (string). The second is the paper-based measure (numeric) and the third is the computer-based measure (numeric).
The measures should arrive to this package adjusted by the pooled standard deviation of the PBA 'and CBA IRT measure per item.}
}
\value{
Returns 1 if initialization is successful.
}
\description{
Initializing function that accepts a dataframe with paper and computer based measures for either item-difficulty,
-discrimination, or -guessing.
}
\examples{
mydata= data.frame(itemID= c("MC1", "MC2", "MC3"),
PBAvals= c(-0.25, 0.00, 0.25), CBAvals= c(-0.30, 0.00, 0.30))
itemmodef (pbcbDF= mydata)

}
